requirements.txt
Django==3.1.4
psycopg2==2.8.6
beautifulsoup4==4.9.3
python-dateutil==2.8.1
requests==2.25.1
